requirements.txt
aiogram>=3.0.0
yt-dlp
python-dotenv
aiohttp
